requirements.txt
matplotlib==3.6.2
nibabel>=3.0.0
numpy==1.23.5
Pillow==9.3.0
tqdm==4.64.1
wandb==0.13.5
